pyproject.toml
[tool.poetry]
name = "stock"
version = "0.1.0"
description = ""
authors = [""]
license = "WTFPL"

[tool.poetry.dependencies]
python = "^3.9"
alpaca_trade_api = "*"
peewee = "*"
python-dotenv = "*"
autobahn = {version = "*", extras = ["encryption", "serialization"]}
invoke = "*"
sty = "*"
schema = "*"
pymonad = "*"
pykka = "*"
janus = "*"
mplfinance = "*"
psycopg2 = "^2.9.3"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"